refactor(datascience): move transformer functions to a file
Refs: #96
